implemented rendering of films list in my library
