+++ src/js/dom/dom-index.js
@@ -1,7 +1,6 @@
 import './scrollToTop';
 import './movie-modal';
 import './input-search';
-import './queue';
-import './watched';
+import './my_library';
 import './create-movie-markup';
 import './create-movie-modal-markup';
