+++ src/js/other/localeStorageServices.js
@@ -25,3 +25,5 @@ export const removeFromStorage = key => {
     console.error(error);
   }
 };
+
+
